requirements.txt
requests
dingtalkchatbot
pyyaml
lxml
feedparser
python-telegram-bot